+++ dev/october-events/includes/Plugin.php
@@ -39,6 +39,7 @@ final class Plugin {
 
         // Services.
         RestApi::get_instance()->init();
+        Cors::init();
         \OE\Planning\Rest::init();
         \OE\Tasks\Rest::init();
         \OE\Volunteers\Rest::init();